osquery/runtime: windows support for TestOsquerySlowStart
Pretty similar to posix, just more powershell.

Also added a healthy check which was missing from the original test.
